pyproject.toml
[project]
name = "assignment1-embeddings"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "branca",
    "dash>=4.0.0",
    "isort>=6.1.0",
    "kaleido>=1.2.0",
    "matplotlib",
    "numba>=0.58.0",
    "numpy<2",
    "optuna>=4.7.0",
    "pandas",
    "plotly>=6.5.2",
    "pyvis",
    "ruff>=0.15.1",
    "scikit-learn",
    "scipy",
    "seaborn",
    "sentence-transformers",
    "setuptools<70",
    "ty>=0.0.17",
    "umap-learn>=0.5.4",
]
